// src/app/admin/admin-shared/admin.interfaces.ts
export interface Admin {
  ACCESS_TOKEN: string;
  ACCT_NO: string;
  ACTIVE: number;
  ADDRESS: string;
  AZURE_ID: string;
  BUSINESS_NAME: string;
  CITY: string;
  CONFIRM_TOKEN: string;
  CONFIRM_TOKEN_EXPIRES: string;
  CONFIRM_TOKEN_TYPE: string;
  CREATED_AT: string;
  EMAIL: string;
  EMAIL_VERIFIED: number;
  EMAIL_VERIFIED_TIME: string;
  FAX: string;
  FIRST_NAME: string;
  LANG: string;
  LAST_NAME: string;
  LOGIN_ATTEMPTS: number;
  PHONE: string;
  POSTAL_CODE: string;
  PROVINCE: string;
  REFRESH_TOKEN: string;
  ROLE: number;
  USER_ID: number;
}

export interface User {
  ACCT_NO: string;
  ACTIVE: number;
  ADDRESS: string;
  AZURE_ID: string;
  BUSINESS_NAME: string;
  CITY: string;
  CREATED_AT: string;
  EMAIL: string;
  EMAIL_VERIFIED: number;
  FAX: string;
  FIRST_NAME: string;
  LANG: string;
  LAST_NAME: string;
  PHONE: string;
  POSTAL_CODE: string;
  PROVINCE: string;
  ROLE: number;
  USER_ID: number;
}

export interface Contact {
  ACCT_NO: string;
  ALT_AREA_CODE: string;
  ALT_E_MAIL: string;
  ALT_PHONE_EXCHANGE: string;
  ALT_PHONE_EXT: string;
  ALT_PHONE_LOC: string;
  ALT_PHONE_NO: string;
  AREA_CODE: string;
  CONTACT_FIRST_NAME: string;
  CONTACT_ID: number
  CONTACT_LAST_NAME: string;
  CONTACT_MIDDLE_NAME: string;
  CONTACT_TITLE: string;
  CORRESPOND_CONTACT: string;
  DEL_FLAG: string;
  DESCR: string;
  EPR_FAX_PHONE_NO: string;
  EPR_PHONE_NO: string;
  E_MAIL: string;
  FAX_AREA_CODE: string;
  FAX_PHONE_EXCHANGE: string;
  FAX_PHONE_NO: string;
  OP_CODE: string;
  PERCONSYSROWID: string;
  PHONE_EXCHANGE: string;
  PHONE_EXT: string;
  PHONE_LOC: string;
  PHONE_NO: string;
  PRIMARY_CONTACT: string;
  REV_DATE: string;
  STMNT_CONTACT: string;
}